actualizaciion migration

+++ main/database/migrations/2023_09_28_101443_create_boards_table.php
@@ -14,7 +14,7 @@ class CreateBoardsTable extends Migration
     public function up()
     {
         Schema::create('boards', function (Blueprint $table) {
-            $table->integer('id');
+            $table->bigIncrements('id');
             $table->string('name_board', 50);
         });
     }
